@@ -5,53 +5,58 @@
   "author": "",
   "private": true,
   "devDependencies": {
-    "babel-core": "^6.8.0",
+    "acorn": "^6.0.5",
+    "ansi-colors": "^3.2.3",
+    "babel-core": "^6.26.3",
     "babel-eslint": "^6.0.2",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.5.0",
     "babelify": "^7.2.0",
     "browserify": "^13.0.0",
     "chai": "^3.5.0",
     "envify": "^3.4.0",
-    "eslint": "^2.7.0",
+    "eslint": "^5.12.0",
     "eslint-plugin-react": "^4.2.3",
-    "gulp": "^3.9.1",
-    "gulp-clean-css": "^2.0.7",
-    "gulp-connect": "^3.2.2",
+    "fancy-log": "^1.3.3",
+    "gulp": "^4.0.0",
+    "gulp-clean-css": "^3.10.0",
+    "gulp-cli": "^2.0.1",
+    "gulp-connect": "^5.7.0",
     "gulp-eslint": "^2.0.0",
     "gulp-gzip": "^1.2.0",
     "gulp-notify": "^2.2.0",
-    "gulp-rename": "^1.2.2",
+    "gulp-rename": "^1.4.0",
     "gulp-replace": "^0.5.4",
-    "gulp-ruby-sass": "^2.0.6",
-    "gulp-sass": "^2.2.0",
+    "gulp-ruby-sass": "^3.0.0",
+    "gulp-sass": "^4.0.2",
     "gulp-tar": "^1.9.0",
     "gulp-uglify": "^1.5.1",
-    "gulp-unzip": "^0.1.3",
-    "gulp-util": "^3.0.7",
-    "gulp-zip": "^3.1.0",
+    "gulp-unzip": "^1.0.0",
+    "gulp-zip": "^4.2.0",
     "merge-stream": "^1.0.0",
     "mkdirp": "^0.5.0",
-    "mocha": "^2.4.5",
-    "react-addons-test-utils": "15.3.2",
+    "mocha": "^5.2.0",
+    "react-test-renderer": "^16.7.0",
     "require-dir": "^0.3.0",
-    "rimraf": "^2.4.4",
-    "run-sequence": "^1.1.5",
-    "sinon": "^1.17.3",
-    "vinyl-source-stream": "^1.1.0",
+    "rimraf": "^2.6.3",
+    "sinon": "^7.2.2",
+    "vinyl-source-stream": "^2.0.0",
     "watchify": "^3.7.0"
   },
   "dependencies": {
     "classnames": "2.2.5",
-    "i18next": "2.5.1",
+    "i18next": "^10.6.0",
+    "i18next-xhr-backend": "^1.5.1",
+    "jquery": "^3.3.1",
+    "jquery-i18next": "^1.2.1",
     "object-assign": "4.1.0",
-    "react": "15.3.2",
-    "react-addons-pure-render-mixin": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-i18next": "1.10.1",
-    "react-router": "2.8.1",
+    "prop-types": "^15.6.2",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
+    "react-i18next": "8.0.7",
+    "react-router": "^3.2.1",
     "react-timer-mixin": "0.13.3",
-    "ua-parser-js": "^0.7.12"
+    "ua-parser-js": "^0.7.18"
   },
   "scripts": {
     "test": "mocha --require babel-core/register test/**/*.js",
