@@ -1,15 +1,15 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import SignInOptionList from './SignInOptionList.js';
+import Translated from '../Translated.js';
+import EidasPrivacyStatementLink from './EidasPrivacyStatementLink.js';
 
-var CountrySelection = React.createClass({
-    propTypes: {
-        availableCountries: React.PropTypes.array.isRequired
-    },
+class CountrySelection extends React.Component {
 
-    getMethodsToShow: function(availableCountries) {
+    getMethodsToShow(countries) {
         let methodsToShow = [];
-        availableCountries.forEach((c) => {
+        countries.forEach((c) => {
             if (c && c.enabled) {
                 methodsToShow.push(c);
             }
@@ -22,18 +22,26 @@ var CountrySelection = React.createClass({
             else return 1;
         });
         return methodsToShow;
-    },
-
-    render: function() {
+    }
+    render() {
         let methodsToShow = this.getMethodsToShow(this.props.availableCountries);
         return (
-            <div className="row">
+            <div className="row" className="sign-in-options">
                 <div className="col-xs-12">
+                    <Translated tag="h2" id="valinta__vaihtoehto__otsikko_valitse_maa" />
                     <SignInOptionList allowedMethods={methodsToShow}/>
+                </div>
+                <hr></hr>
+                <div className="col-xs-12">
+                    <EidasPrivacyStatementLink/>
                 </div>
             </div>
         );
-    },
-});
+    }
+}
+
+CountrySelection.propTypes = {
+    availableCountries: PropTypes.array.isRequired
+};
 
 export default CountrySelection;
