@@ -1,22 +1,21 @@
 ﻿import React from 'react';
 
-//import ReactBase from '../ReactBase';
 import TranslatedImage from '../TranslatedImage.js';
 import Translated from '../Translated.js';
 
-let Header =  React.createClass({
-    render: function() {
+class Header extends React.Component {
+    render() {
         return (
           <header id="page-header" role="banner">
               <div id="header-content" className={'container'}>
-                  <h1 id="suomi.fi-tunnistaminen" className="site-logo">
+                  <div id="suomi.fi-tunnistaminen" className="site-logo">
                       <TranslatedImage srcKey="header__logo" altKey="header__suomifi-tunnistaminen" />
                       <Translated tag="span" id="header__suomifi-tunnistaminen" className="visuallyhidden" />
-                  </h1>
+                  </div>
               </div>
           </header>
         );
     }
-});
+}
 
 export default Header;
