@@ -1,21 +1,24 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-var LanguageButton = React.createClass({
-    contextTypes: {
-        queryParams: React.PropTypes.object
-    },
-    propTypes: {
-        buttonLanguage: React.PropTypes.string.isRequired,
-        selected: React.PropTypes.bool,
-        text: React.PropTypes.string.isRequired
-    },
-    render: function() {
+class LanguageButton extends React.Component {
+    render() {
         return (
             <a className={this.props.selected ? 'selected' : ''}>
                 <span className={this.props.selected ? 'selected' : ''}>{this.props.text}</span>
             </a>
         );
     }
-});
+}
+
+LanguageButton.contextTypes = {
+    queryParams: PropTypes.object
+};
+
+LanguageButton.propTypes = {
+    buttonLanguage: PropTypes.string.isRequired,
+    selected: PropTypes.bool,
+    text: PropTypes.string.isRequired
+};
 
 export default LanguageButton;
