@@ -16,7 +16,7 @@
 
 var gulp = require('gulp');
 var requireDir = require('require-dir');
-var runSequence = require('run-sequence');
+var log = require('fancy-log');
 
 // globals
 global.root = "dev";
@@ -26,19 +26,15 @@ global.pagesDir = 'sivut';
 // Require all tasks in gulp/tasks
 requireDir('./tasks', { recurse: true });
 
-//----dist-build-stuff----:
-// (Keeping the eslint + minify up here in gulpfile. For now that is ok)
-gulp.task('dist-build', function (callback) {
-    runSequence('dist:clean', 'dist:init',
-        'dist:build-dist-files',
-        'js-lint-n-min', callback);
-});
 
+//----dist-build-stuff----:
 // Makes sense to run this task after the other dist-build tasks;
 // handled by the calling task ('dist-build') in its runSequence.
-gulp.task('js-lint-n-min', function (done) {
-    runSequence('eslint', 'minify', done);
-});
+gulp.task('js-lint-n-min', gulp.series('eslint', 'minify'));
+
+// (Keeping the eslint + minify up here in gulpfile. For now that is ok)
+const distBuildSeries = gulp.series('dist:clean', 'dist:init','dist:build-dist-files','js-lint-n-min');
+gulp.task('dist-build', (done) => { distBuildSeries(done) });
 //----dist-build-stuff----}
 
 
@@ -47,32 +43,22 @@ gulp.task('setWatch', function () {
     global.isWatching = true;
 });
 
-gulp.task('dev-build', function () {
-    return runSequence('dev:init',
-        'eslint', 'browserify',
-        'dev:build-dev-files');
-});
+gulp.task('dev-build', gulp.series('dev:init', 'eslint', 'browserify','dev:build-dev-files'));
 
-// build files needed by non-react pages
-gulp.task('dev-build-static-resources', function () {
-    return runSequence('dev:init', "dev:fonts", "dev:images", "dev:js", "dev:sass");
-});
-
-gulp.task('watch', function () {
-    runSequence('dev-build', 'serve');
+gulp.task('watch-all', function() { 
     gulp.watch([
-        './*.html',
-        './font/**/*',
-        './img/**/*',
-        './images/**/*',
-        './locale/**/*',
-        './sass/**/*.scss',
-        './css/**/*.css',
-        './fixture/**/*',
-        './src/**/*.js',
-        './js/**/*.js'], function() {
-        runSequence('dev:init',
-        'eslint', 'browserify', 'dev:build-dev-files');
-       });
+    './*.html',
+    './font/**/*',
+    './img/**/*',
+    './images/**/*',
+    './locale/**/*',
+    './sass/**/*.scss',
+    './css/**/*.css',
+    './fixture/**/*',
+    './src/**/*.js',
+    './js/**/*.js'], gulp.series('dev:init', 'eslint', 'browserify', 'dev:build-dev-files'))
 });
+
+gulp.task('watch', gulp.series('dev-build', gulp.parallel('serve', 'watch-all')));
+
 //----dev-build-stuff----}
