@@ -1,11 +1,8 @@
-import { expect } from 'chai';
-import { sandbox as sinonSandbox, stub, spy, match } from 'sinon';
+import {expect} from 'chai';
+import sinon from 'sinon';
 import i18n from 'i18next';
-
-
-
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
+import ShallowRenderer from 'react-test-renderer/shallow';
 
 import TranslatedLink from '../../src/TranslatedLink.js';
 
@@ -14,18 +11,18 @@ describe('TranslatedLink', function() {
         let sandbox;
         let translate;
         beforeEach(function() {
-            sandbox = sinonSandbox.create();
+            sandbox = sinon.createSandbox();
             translate = sandbox.stub(i18n, 't');
             translate.withArgs('testLinkId').returns('testLinkTranslation');
             translate.withArgs('testContentId').returns('testContentTranslation');
         });
 
-        afterEach(function () {
+        afterEach(function() {
             sandbox.restore();
         });
 
         it('should show the a-element', function() {
-            let renderer = ReactTestUtils.createRenderer();
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
@@ -35,17 +32,17 @@ describe('TranslatedLink', function() {
         });
 
         it('should have data-i18n attribute', function() {
-            let renderer = ReactTestUtils.createRenderer();
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
 
             let result = renderer.getRenderOutput();
-            expect(result.props["data-i18n"]).to.equal('testContentId');
+            expect(result.props['data-i18n']).to.equal('testContentId');
         });
 
         it('should get translated link', function() {
-            let renderer = ReactTestUtils.createRenderer();
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
@@ -55,7 +52,7 @@ describe('TranslatedLink', function() {
         });
 
         it('should get translated content', function() {
-            let renderer = ReactTestUtils.createRenderer();
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
@@ -66,21 +63,21 @@ describe('TranslatedLink', function() {
     });
     describe('TranslatedLink fallback', function() {
         let sandbox;
-        let translate
+        let translate;
         beforeEach(function() {
-            sandbox = sinonSandbox.create();
+            sandbox = sinon.createSandbox();
             translate = sandbox.stub(i18n, 't');
-            i18n.options = { 'fallbackLng': ['fi'] };
+            i18n.options = {'fallbackLng': ['fi']};
         });
 
-        afterEach(function () {
+        afterEach(function() {
             sandbox.restore();
         });
 
         it('should get use finnish fallback link if no localised link available', function() {
-            translate.withArgs('testLinkId', match.has('lng', 'fi')).returns('fallbackLinkTranslation')
-                     .withArgs('testLinkId', match.typeOf('undefined')).returns(null);
-            let renderer = ReactTestUtils.createRenderer();
+            translate.withArgs('testLinkId', sinon.match.has('lng', 'fi')).returns('fallbackLinkTranslation')
+                     .withArgs('testLinkId', sinon.match.typeOf('undefined')).returns(null);
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
