@@ -1,11 +1,8 @@
-import { expect } from 'chai';
-import { sandbox as sinonSandbox, stub, spy, match } from 'sinon';
+import {expect} from 'chai';
+import sinon from 'sinon';
 import i18n from 'i18next';
-
-
-
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
+import ShallowRenderer from 'react-test-renderer/shallow';
 
 import TranslatedPaddedLink from '../../src/TranslatedPaddedLink.js';
 
@@ -14,18 +11,18 @@ describe('TranslatedPaddedLink', function() {
         let sandbox;
         let translate;
         beforeEach(function() {
-            sandbox = sinonSandbox.create();
+            sandbox = sinon.createSandbox();
             translate = sandbox.stub(i18n, 't');
             translate.withArgs('testLinkId').returns('testLinkTranslation');
             translate.withArgs('testContentId').returns('testContentTranslation');
         });
 
-        afterEach(function () {
+        afterEach(function() {
             sandbox.restore();
         });
 
         it('should show a span element', function() {
-            let renderer = ReactTestUtils.createRenderer();
+            let renderer = ShallowRenderer.createRenderer();
             renderer.render(
                 <TranslatedPaddedLink id="testId" className="testClass" link_i18n_id="testLinkId" content_i18n_id="testContentId"/>
             );
