@@ -1,8 +1,8 @@
-import { expect } from 'chai';
-import { useFakeXMLHttpRequest, mock, spy } from 'sinon';
+import {expect} from 'chai';
+import sinon from 'sinon';
 
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
+import ShallowRenderer from 'react-test-renderer/shallow';
 
 import FeedbackPage from '../../src/feedbackPage/FeedbackPage.js';
 import FeedbackForm from '../../src/feedbackPage/FeedbackForm.js';
@@ -10,9 +10,9 @@ import FeedbackForm from '../../src/feedbackPage/FeedbackForm.js';
 describe('FeedbackPage', function() {
 
     it('should show the form', function() {
-        let renderer = ReactTestUtils.createRenderer();
+        let renderer = ShallowRenderer.createRenderer();
         let router = {
-            push: spy()
+            push: sinon.spy()
         };
         renderer.render(
             <FeedbackPage>
